# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AMHARC Match Capture — Local Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches:
    get:
      operationId: getMatches
      tags: [matches]
      summary: List all matches
      responses:
        "200":
          description: Match list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
    post:
      operationId: createMatch
      tags: [matches]
      summary: Create a new match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchInput"
      responses:
        "201":
          description: Match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{matchId}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get match details
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "404":
          description: Not found
    put:
      operationId: updateMatch
      tags: [matches]
      summary: Update match details
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{matchId}/start:
    post:
      operationId: startMatch
      tags: [matches]
      summary: Start a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResult"
  /matches/{matchId}/stop:
    post:
      operationId: stopMatch
      tags: [matches]
      summary: Stop a match
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationResult"
  /matches/{matchId}/clock/start:
    post:
      operationId: startMatchClock
      tags: [matches]
      summary: Start match clock
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Clock started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClockState"
  /matches/{matchId}/clock/pause:
    post:
      operationId: pauseMatchClock
      tags: [matches]
      summary: Pause match clock
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Clock paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClockState"
  /matches/{matchId}/clock/resume:
    post:
      operationId: resumeMatchClock
      tags: [matches]
      summary: Resume match clock
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Clock resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClockState"
  /matches/{matchId}/clock/correct:
    post:
      operationId: correctMatchClock
      tags: [matches]
      summary: Manually correct match clock
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockCorrection"
      responses:
        "200":
          description: Clock corrected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClockState"
  /matches/{matchId}/clock:
    get:
      operationId: getMatchClock
      tags: [matches]
      summary: Get match clock state
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Clock state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClockState"
  /matches/{matchId}/score:
    get:
      operationId: getMatchScore
      tags: [matches]
      summary: Get current match score
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Score state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreState"
    put:
      operationId: updateMatchScore
      tags: [matches]
      summary: Update match score
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreUpdate"
      responses:
        "200":
          description: Score updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScoreState"
...
components:
  schemas:
...
    Match:
      type: object
      required: [matchId, humanId, sport, competition, season, homeTeam, awayTeam, date, status, createdAt]
      properties:
        matchId:
          type: string
        humanId:
          type: string
        sport:
          type: string
          enum: [gaelic-football, hurling, ladies-football, camogie]
        competition:
          type: string
        season:
          type: string
        round:
          type: ["string", "null"]
        date:
          type: string
          format: date
        venue:
          type: ["string", "null"]
        homeTeam:
          type: string
        awayTeam:
          type: string
        homeTeamShort:
          type: ["string", "null"]
        awayTeamShort:
          type: ["string", "null"]
        homeTeamColour:
          type: ["string", "null"]
        awayTeamColour:
          type: ["string", "null"]
        operator:
          type: ["string", "null"]
        scheduledStart:
          type: ["string", "null"]
          format: date-time
        periodStructure:
          type: string
          enum: [halves, quarters, custom]
        expectedDurationMinutes:
          type: ["integer", "null"]
        recordingDirectory:
          type: ["string", "null"]
        cameraId:
          type: ["string", "null"]
        streamProfile:
          type: ["string", "null"]
        overlayTemplate:
          type: ["string", "null"]
        streamDestination:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [setup, ready, active, halftime, complete, cancelled]
        currentPeriod:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    MatchInput:
      type: object
      required: [sport, competition, season, homeTeam, awayTeam, date]
      properties:
        sport:
          type: string
          enum: [gaelic-football, hurling, ladies-football, camogie]
        competition:
          type: string
        season:
          type: string
        round:
          type: string
        date:
          type: string
          format: date
        venue:
          type: string
        homeTeam:
          type: string
        awayTeam:
          type: string
        homeTeamShort:
          type: string
        awayTeamShort:
          type: string
        homeTeamColour:
          type: string
        awayTeamColour:
          type: string
        operator:
          type: string
        scheduledStart:
          type: string
          format: date-time
        periodStructure:
          type: string
          enum: [halves, quarters, custom]
        expectedDurationMinutes:
          type: integer
        recordingDirectory:
          type: string
        cameraId:
          type: string
        streamProfile:
          type: string
        overlayTemplate:
          type: string
        streamDestination:
          type: string
        notes:
          type: string
    MatchUpdate:
      type: object
      properties:
        competition:
          type: string
        round:
          type: string
        venue:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [setup, ready, active, halftime, complete, cancelled]
    ClockState:
      type: object
      required: [matchId, matchClockSeconds, recordingElapsedSeconds, isRunning, currentPeriod, clockMode]
      properties:
        matchId:
          type: string
        matchClockSeconds:
          type: integer
        recordingElapsedSeconds:
          type: integer
        isRunning:
          type: boolean
        currentPeriod:
          type: integer
        clockMode:
          type: string
          enum: [count-up, count-down]
        updatedAt:
          type: string
          format: date-time
    ClockCorrection:
      type: object
      required: [matchClockSeconds]
      properties:
        matchClockSeconds:
          type: integer
        reason:
          type: string
    ScoreState:
      type: object
      required: [matchId, homeGoals, homePoints, awayGoals, awayPoints, homeTotal, awayTotal]
      properties:
        matchId:
          type: string
        homeGoals:
          type: integer
        homePoints:
          type: integer
        awayGoals:
          type: integer
        awayPoints:
          type: integer
        homeTotal:
          type: integer
        awayTotal:
          type: integer
        updatedAt:
          type: ["string", "null"]
          format: date-time
    ScoreUpdate:
      type: object
      required: [team, scoreType, delta]
      properties:
        team:
          type: string
          enum: [home, away]
        scoreType:
          type: string
          enum: [goal, point, two-point]
        delta:
          type: integer
        reason:
          type: ["string", "null"]
...
    OperationResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
        error:
          type: ["string", "null"]