# lib/api-zod/src/generated/types/clockCorrection.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AMHARC Match Capture — Local Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ClockCorrection {
  matchClockSeconds: number;
  reason?: string;
}