# lib/api-zod/src/generated/types/matchInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AMHARC Match Capture — Local Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchInputPeriodStructure } from './matchInputPeriodStructure';
import type { MatchInputSport } from './matchInputSport';

export interface MatchInput {
  sport: MatchInputSport;
  competition: string;
  season: string;
  round?: string;
  date: Date;
  venue?: string;
  homeTeam: string;
  awayTeam: string;
  homeTeamShort?: string;
  awayTeamShort?: string;
  homeTeamColour?: string;
  awayTeamColour?: string;
  operator?: string;
  scheduledStart?: Date;
  periodStructure?: MatchInputPeriodStructure;
  expectedDurationMinutes?: number;
  recordingDirectory?: string;
  cameraId?: string;
  streamProfile?: string;
  overlayTemplate?: string;
  streamDestination?: string;
  notes?: string;
}