# lib/api-zod/src/generated/types/matchUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AMHARC Match Capture — Local Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchUpdateStatus } from './matchUpdateStatus';

export interface MatchUpdate {
  competition?: string;
  round?: string;
  venue?: string;
  notes?: string;
  status?: MatchUpdateStatus;
}