# lib/api-zod/src/generated/types/scoreState.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * AMHARC Match Capture — Local Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreState {
  matchId: string;
  homeGoals: number;
  homePoints: number;
  awayGoals: number;
  awayPoints: number;
  homeTotal: number;
  awayTotal: number;
  /** @nullable */
  updatedAt?: Date | null;
}